+++ sql/sqlfs/README.md
@@ -1,15 +1,15 @@
-# `sqlfile`
+# `sqlfs`
 
-The package `sqlfile` provides an `io.ReadCloser` and an `io.WriteCloser` that treats a SQL database a filesystem, where each table in the database is like a file.  The schema of the table is very simple -- it has only one column of BLOB type. All the rows consist the storage.
+The package `sqlfs` provides an `io.ReadCloser` and an `io.WriteCloser` that treats a SQL database a filesystem, where each table in the database is like a file.  The schema of the table is very simple -- it has only one column of BLOB type. All the rows consist the storage.
 
-`sqlfile` provides the following features.
+`sqlfs` provides the following features.
 
 ## Create a file
 
 To create a table named "hello" in a database "mydb" for writing, we can call `Create`.
 
 ```go
-f, e := sqlfile.Create(db, "mydb.hello")
+f, e := sqlfs.Create(db, "mydb.hello")
 f.Write([]byte("hello world!\n"))
 f.Close()
 ```
@@ -19,7 +19,7 @@ where `db` comes from a call to `database/sql.Open`.
 ## Append to a file
 
 ```go
-f, e := sqlfile.Append(db, "mydb.hello")
+f, e := sqlfs.Append(db, "mydb.hello")
 f.Write([]byte("hello world!\n")
 f.Close()
 ```
@@ -27,7 +27,7 @@ f.Close()
 ## Read from a file
 
 ```go
-f, e := sqlfile.Open(db, "mydb.hello")
+f, e := sqlfs.Open(db, "mydb.hello")
 buf := make([]byte, 1024)
 f.Read(buf)
 f.Close()
@@ -47,10 +47,10 @@ HasTable(db, "mydb.hello")
 
 ## Other I/O operations
 
-Feel free to use standard packages `io`, `ioutil`, etc with `sqlfile`.  For example, we can call `io.Copy` to copy everything from the standard input to a table.
+Feel free to use standard packages `io`, `ioutil`, etc with `sqlfs`.  For example, we can call `io.Copy` to copy everything from the standard input to a table.
 
 ```go
-f, e := sqlfile.Create(db, "mydb.hello")
+f, e := sqlfs.Create(db, "mydb.hello")
 io.Copy(f, os.Stdin)
 f.Close()
 ```
